11-05-2022 @ 9:08 AM
I commented out the bicep script resource and all associated variables so that I can test to see if this bicep file deploys properly. I also changed the resource group that is defined in the yaml file and the name of the deployment that is defined in the resource file as well.