11-05-2022 @ 9:58 AM
My fix for string interpolation was a success so I proceeded with enabling all powershell script variables and resources in both the bicep file and the yaml file. I am now going to deploy the bicep file again and see if the powershell script works without any interpolation errors and I'll make certain that the powershell script is successfully by checking it's output for items it's supposed to query from the storage account container "images".