11-06-2022 @ 8:47 AM
I found that if you assign the AD group object ID directly to the New-AzRoleAssignment it fixes it without issue. This is not efficient though because we can not use the name of the group to find the ID due to Get-AzGroup not working due to not having enough privileges. Right now I'm adding directory access tot he Microsoft Graph API permissions and seeing if Get-AzADGroup will work with Connect-AzAccount -Identity.